temp stuff to work on other machine
