--- src/Entity/ReadWritePlayer/ReadPlayerFromJson.java
@@ -1,6 +1,0 @@
-package Entity.ReadWritePlayer;
-
-public class ReadPlayerFromJson {
-
-
-}
