+++ src/ReadAndWrite/ReadWritePlayer/ReadPlayerFromJson.java
@@ -1,0 +1,13 @@
+package ReadAndWrite.ReadWritePlayer;
+
+import com.google.gson.Gson;
+import com.google.gson.JsonObject;
+
+import java.io.FileReader;
+import java.io.IOException;
+
+public class ReadPlayerFromJson {
+
+
+
+}
